uitester/requirements.txt
dapr
dapr-ext-grpc
chainlit
pydantic==2.10.1
azure-identity
azure-servicebus
azure-storage-blob